+++ NEWS.md
@@ -1,12 +1,16 @@
 # x3ptools (development version)
 
-# x3ptools 0.0.3.9000
+# x3ptools 0.0.4.9000
+
+
+# x3ptools 0.0.4
 
 ## major functionality 
 
 * `tmd_to_x3p` allows an import of a file in TMD format (natively supported by GelSight instruments) to x3p
 * `x3p_bin_stripes` color-shade the mask by row or column-based gradient
 * `x3p_shade_mask` color-shade the mask of an x3p by its surface profile
+* `x3p_circle_select` interactively select a circle on an x3p surface by selecting three points
 
 
 ## minor functionality